src/subTestUtils: add tests for escapeSubTestName
